+++ src/components/DailyInfo/DailyInfo.module.css
@@ -1,0 +1,52 @@
+.waterListBlock {
+  padding-bottom: 0px;
+}
+
+.waterListBlockHead {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  height: 32px;
+  padding-left: 16px;
+}
+
+.wrapper {
+  position: relative;
+  height: 98px;
+}
+
+@media screen and (min-width: 768px) {
+  .wrapper {
+    height: 118px;
+  }
+}
+
+.waterListError {
+  position: relative;
+  font-size: 16px;
+  text-align: center;
+}
+
+.errorIcon {
+  position: absolute;
+  top: 40px;
+  left: 50%;
+  width: 50px;
+  height: 50px;
+  color: rgb(246, 39, 39);
+
+  transform: translateX(-50%);
+}
+
+@media screen and (min-width: 768px) {
+  .errorIcon {
+    width: 80px;
+    height: 80px;
+  }
+}
+
+@media screen and (min-width: 768px) {
+  .waterListError {
+    font-size: 18px;
+  }
+}
