+++ src/components/UserPanelAvatar/UserPanelAvatar.jsx
