+++ src/components/UserPanelAvatar/UserPanelAvatar.module.css
