--- src/components/WaterDetailedInfo/DailyInfo/DailyInfo.module.css
@@ -1,103 +1,0 @@
-/* .waterListBlock {
-  padding-bottom: 40px;
-}
-
-.waterListBlockHead {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-bottom: 24px;
-}
-
-.wrapper {
-  position: relative;
-  height: 98px;
-}
-
-@media screen and (min-width: 768px) {
-  .wrapper {
-    height: 118px;
-  }
-}
-
-.waterListError {
-  position: relative;
-  font-size: 16px;
-  text-align: center;
-}
-
-.errorIcon {
-  position: absolute;
-  top: 40px;
-  left: 50%;
-  width: 50px;
-  height: 50px;
-  color: rgb(246, 39, 39);
-
-  transform: translateX(-50%);
-}
-
-@media screen and (min-width: 768px) {
-  .errorIcon {
-    width: 80px;
-    height: 80px;
-  }
-}
-
-@media screen and (min-width: 768px) {
-  .waterListError {
-    font-size: 18px;
-  }
-} */
-/* =========================================== */
-.waterListBlock {
-  padding-bottom: 40px;
-}
-
-.waterListBlockHead {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-bottom: 24px;
-}
-
-.wrapper {
-  position: relative;
-  height: 98px;
-}
-
-@media screen and (min-width: 768px) {
-  .wrapper {
-    height: 118px;
-  }
-}
-
-.waterListError {
-  position: relative;
-  font-size: 16px;
-  text-align: center;
-}
-
-.errorIcon {
-  position: absolute;
-  top: 40px;
-  left: 50%;
-  width: 50px;
-  height: 50px;
-  color: rgb(246, 39, 39);
-
-  transform: translateX(-50%);
-}
-
-@media screen and (min-width: 768px) {
-  .errorIcon {
-    width: 80px;
-    height: 80px;
-  }
-}
-
-@media screen and (min-width: 768px) {
-  .waterListError {
-    font-size: 18px;
-  }
-}
