+++ src/components/WaterDetailedInfo/WaterDetailedInfo.jsx
@@ -1,5 +1,5 @@
-import { UserPanel } from './UserPanel/UserPanel';
-import DailyInfo from './DailyInfo/DailyInfo';
+import { UserPanel } from '../UserPanel/UserPanel';
+import DailyInfo from '../DailyInfo/DailyInfo';
 import s from './WaterDetailedInfo.module.css';
 
 const WaterDetailedInfo = () => {
