+++ src/components/WaterDetailedInfo/WaterDetailedInfo.module.css
@@ -11,6 +11,8 @@
 @media screen and (min-width: 768px) {
   .waterDetailedInfo {
     padding: 32px;
+    padding-top: 41px;
+    padding-bottom: 0;
     width: 704px;
   }
 }
