+++ src/components/WaterList/WaterList.jsx
@@ -1,0 +1,12 @@
+import WaterItem from '../Wate';
+// import s from './WaterList.module.css';
+
+const WaterList = () => {
+  return (
+    <div>
+      <WaterItem />
+    </div>
+  );
+};
+
+export default WaterList;
